# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Personal household budget API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiMessage {
  error: string;
}

export interface Settings {
  selectedMonth: string;
  monthStartDay: number;
  checkingBuffer: number;
  debtStrategy: string;
}

export interface SettingsUpdate {
  selectedMonth?: string;
  /**
     * @minimum 1
     * @maximum 28
     */
  monthStartDay?: number;
  checkingBuffer?: number;
  debtStrategy?: string;
}

export interface IncomeSource {
  id: number;
  name: string;
  /** @nullable */
  owner?: string | null;
  frequency: string;
  netAmount: number;
  monthlyEquivalent: number;
  /** @nullable */
  notes?: string | null;
}

export interface IncomeSourceInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  owner?: string | null;
  frequency: string;
  netAmount: number;
  monthlyEquivalent: number;
  /** @nullable */
  notes?: string | null;
}

export interface IncomeSourceUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  owner?: string | null;
  frequency?: string;
  netAmount?: number;
  monthlyEquivalent?: number;
  /** @nullable */
  notes?: string | null;
}

export interface Category {
  id: number;
  name: string;
  subcategories: string[];
  sortOrder: number;
}

export interface CategoryInput {
  /** @minLength 1 */
  name: string;
  subcategories?: string[];
  sortOrder?: number;
}

export interface CategoryUpdate {
  /** @minLength 1 */
  name?: string;
  subcategories?: string[];
  sortOrder?: number;
}

export interface PlanLine {
  id: number;
  category: string;
  subcategory: string;
  planned: number;
  /** @nullable */
  priority?: string | null;
  /** @nullable */
  fixedVariable?: string | null;
  /** @nullable */
  dueDay?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface PlanLineInput {
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  subcategory: string;
  planned: number;
  /** @nullable */
  priority?: string | null;
  /** @nullable */
  fixedVariable?: string | null;
  /** @nullable */
  dueDay?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface PlanLineUpdate {
  /** @minLength 1 */
  category?: string;
  /** @minLength 1 */
  subcategory?: string;
  planned?: number;
  /** @nullable */
  priority?: string | null;
  /** @nullable */
  fixedVariable?: string | null;
  /** @nullable */
  dueDay?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface Transaction {
  id: number;
  date: string;
  description: string;
  /** @nullable */
  originalDescription?: string | null;
  /** @nullable */
  bankCategory?: string | null;
  amount: number;
  status: string;
  /** @nullable */
  account?: string | null;
  source: string;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  subcategory?: string | null;
  include: boolean;
  month: string;
  /** @nullable */
  note?: string | null;
  needsReview: boolean;
}

export interface TransactionInput {
  date: string;
  /** @minLength 1 */
  description: string;
  amount: number;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  subcategory?: string | null;
  /** @nullable */
  account?: string | null;
  /** @nullable */
  note?: string | null;
}

export interface TransactionUpdate {
  /** @nullable */
  category?: string | null;
  /** @nullable */
  subcategory?: string | null;
  include?: boolean;
  /** @nullable */
  note?: string | null;
  saveRule?: boolean;
  rulePattern?: string;
}

export interface CsvImport {
  /** @minLength 1 */
  csvContent: string;
  /** @nullable */
  fileName?: string | null;
}

export interface ImportResult {
  totalRows: number;
  added: number;
  duplicates: number;
  pendingReplaced: number;
  autoCategorized: number;
  needsReview: number;
}

export interface Rule {
  id: number;
  pattern: string;
  category: string;
  subcategory: string;
}

export interface RuleInput {
  /** @minLength 2 */
  pattern: string;
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  subcategory: string;
}

export interface CategoryActual {
  category: string;
  planned: number;
  actual: number;
  remaining: number;
}

export interface DashboardSummary {
  month: string;
  incomePlanned: number;
  incomeActual: number;
  plannedExpenses: number;
  actualExpenses: number;
  remaining: number;
  cashFlow: number;
  reviewCount: number;
  pendingCount: number;
  byCategory: CategoryActual[];
}

export interface ReconciliationRow {
  category: string;
  manualTotal: number;
  bankTotal: number;
  difference: number;
  status: string;
}

export type ListTransactionsParams = {
month?: string;
needsReview?: boolean;
source?: ListTransactionsSource;
};

export type ListTransactionsSource = typeof ListTransactionsSource[keyof typeof ListTransactionsSource];


export const ListTransactionsSource = {
  bank: 'bank',
  manual: 'manual',
} as const;

export type GetDashboardParams = {
month?: string;
};

export type GetReconciliationParams = {
month?: string;
};