# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Personal household budget API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List transactions, optionally filtered
      parameters:
        - name: month
          in: query
          required: false
          schema: { type: string }
        - name: needsReview
          in: query
          required: false
          schema: { type: boolean }
        - name: source
          in: query
          required: false
          schema: { type: string, enum: [bank, manual] }
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Add a manual transaction (cash spending etc.)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}:
    patch:
      operationId: updateTransaction
      tags: [transactions]
      summary: Categorize or edit a transaction; optionally save the choice as a rule
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      summary: Delete a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /import:
    post:
      operationId: importCsv
      tags: [import]
      summary: Import a bank CSV download (deduped, auto-categorized)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CsvImport"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
        "400":
          description: Invalid CSV
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
components:
  schemas:
...
    ApiMessage:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    Transaction:
      type: object
      required: [id, date, description, amount, status, source, include, month, needsReview]
      properties:
        id: { type: number }
        date: { type: string }
        description: { type: string }
        originalDescription: { type: ["string", "null"] }
        bankCategory: { type: ["string", "null"] }
        amount: { type: number }
        status: { type: string }
        account: { type: ["string", "null"] }
        source: { type: string }
        category: { type: ["string", "null"] }
        subcategory: { type: ["string", "null"] }
        include: { type: boolean }
        month: { type: string }
        note: { type: ["string", "null"] }
        needsReview: { type: boolean }
    TransactionInput:
      type: object
      required: [date, description, amount]
      properties:
        date: { type: string }
        description: { type: string, minLength: 1 }
        amount: { type: number }
        category: { type: ["string", "null"] }
        subcategory: { type: ["string", "null"] }
        account: { type: ["string", "null"] }
        note: { type: ["string", "null"] }
    TransactionUpdate:
      type: object
      properties:
        category: { type: ["string", "null"] }
        subcategory: { type: ["string", "null"] }
        include: { type: boolean }
        note: { type: ["string", "null"] }
        saveRule: { type: boolean }
        rulePattern: { type: string }
    CsvImport:
      type: object
      required: [csvContent]
      properties:
        csvContent: { type: string, minLength: 1 }
        fileName: { type: ["string", "null"] }
    ImportResult:
      type: object
      required: [totalRows, added, duplicates, pendingReplaced, autoCategorized, needsReview]
      properties:
        totalRows: { type: number }
        added: { type: number }
        duplicates: { type: number }
        pendingReplaced: { type: number }
        autoCategorized: { type: number }
        needsReview: { type: number }
...