# lib/api-zod/src/generated/types/csvImport.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Personal household budget API
 * OpenAPI spec version: 0.1.0
 */

export interface CsvImport {
  /** @minLength 1 */
  csvContent: string;
  /** @nullable */
  fileName?: string | null;
}