# lib/api-zod/src/generated/types/importResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Personal household budget API
 * OpenAPI spec version: 0.1.0
 */

export interface ImportResult {
  totalRows: number;
  added: number;
  duplicates: number;
  pendingReplaced: number;
  autoCategorized: number;
  needsReview: number;
}