# lib/api-zod/src/generated/types/transactionUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Personal household budget API
 * OpenAPI spec version: 0.1.0
 */

export interface TransactionUpdate {
  /** @nullable */
  category?: string | null;
  /** @nullable */
  subcategory?: string | null;
  include?: boolean;
  /** @nullable */
  note?: string | null;
  saveRule?: boolean;
  rulePattern?: string;
}